+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - es6 `be.int16Array`, Int16Array object assertion
 - es6 `be.uint16Array`, Uint16Array object assertion
 - es6 `be.int32Array`, Int32Array object assertion
+- es6 `be.uint32Array`, Uint32Array object assertion
 
 ## [1.4.1] - 2017-07-10